chore(pr): solve pr comments
